// src/api/generated/model/catalogPage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LivreMente.Api
 * OpenAPI spec version: 1.0
 */
import type { CatalogCounts } from './catalogCounts';
import type { Publication } from './publication';

export interface CatalogPage {
  items: Publication[];
  page: number;
  pageSize: number;
  total: number;
  totalPages: number;
  counts: CatalogCounts;
}