// src/api/generated/model/publication.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LivreMente.Api
 * OpenAPI spec version: 1.0
 */
import type { PublicationType } from './publicationType';
import type { ReadingLanguage } from './readingLanguage';

export interface Publication {
  id: string;
  title: string;
  author: string;
  type: PublicationType;
  language: ReadingLanguage;
  /**
     * Slug de gênero (livro) ou área de conhecimento (artigo).
     * @nullable
     */
  genre?: string | null;
  /** @nullable */
  year?: number | null;
  /**
     * EPUB, PDF...
     * @nullable
     */
  format?: string | null;
  /**
     * Project Gutemberg | arXiv.
     * @nullable
     */
  source?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverUrl?: string | null;
}